tec: Provide a basic rework of the Minz framework
I developped Minz a few years ago when I was young and innocent, mainly
to learn how to develop a web framework. I was highly inspired by Zend,
but I wasn't able to identify a lot of mistakes I made. I didn't expect
Minz to be used years later in FreshRSS, with a lot of people using it.

I learned a lot these last years, first by trying to redevelop Minz in a
simpler way (I never finished), then by working with Ruby on Rails and
Django. Also because I'm older and more experienced.

I wanted a small new project in PHP to improve my skills and to practice
unit tests. I thought it was a good opportunity to redevelop Minz from
scratch, once again :). It will be the base for a more complex project.

I'm very happy with the actual global design and architecture of Minz.
It is still extremely (too) simple, but I think it's a very good start
for the upcoming work on Webubbub!

References:

- https://github.com/marienfressinaud/MINZ
- https://github.com/FreshRSS/FreshRSS