tec: Reorganize Minz files
I made a mistake when I wrote the tests suite for Minz. I considered
Minz was part of the application, while it is separated code. The
configuration file cannot be the same since I'm testing two different
apps: a very simple one when testing Minz, the real one when testing the
application.

For this reason, I moved the tests related to Minz, next to their source
code (under `./src/lib/Minz`). The application tests will be placed under
the `./tests` folder.

I added a `make` target to test Minz exclusively.