tec: Reorganize application structure
I realized that it was better to store actions next to the view files
and other related controller files. I wanted a structure as the
following:

    src/
    └── rabbits
        ├── views
        │   └── items.txt
        └── controller.php

It's kind of a big change, but I'm really sure it will improve the way
of organizing the applications. And it's better now than too late!