tec: Provide database access classes
Accessing a database might be complicated without proper tooling.

The Database class aims to ease the initialization of a PDO object,
using the configuration to automatically initialized a database access.

The DatabaseModel class is designed to help developers with some often
used basic methods (create, read, update and delete), automatically
generated from a set of options.