fix: Check pattern element exist in Router
When exploding the `/rabbit` string on `/`, it returns an array of two
elements where the first is an empty string.

A bug happened when we tried to access the first character of this empty
string.

A first approach would be to apply a `array_filter` on the two arrays to
get rid of those empty values. The problem is that paths with two
following slashes would match with a pattern with only one slash. I'm
not sure yet that it is something I want (e.g. `/rabbits//details` would
match with `/rabbits/details`).

Since I'm not sure of the best behaviour, I prefer to keep the actual
one and just check that the string isn't empty.

The bug was already tested by previous tests, but the error reporting
wasn't activated. It will be done in a following commit.