tec: Initialize the environment
I had some bugs which was failing silently. I realized that when an
obvious one wasn't reported. I realized that I was using a default
configuration.

This Environment class is destined to configure the global env of the
application. It doesn't do a lot for now, but there might be some other
use for it!