tec: Make DatabaseModel attributes protected
I had to create a new method to allow a custom search, in a class
inheriting from DatabaseModel. I wanted it to be as generic as possible
and had to access the table name. I quickly realised that it was a
private attribute, and I wasn't able to use it.

I made the primary_key_name and properties protected as well for the
same reason, even if I didn't have any use of them yet.