tec: Catch errors at a higher level in Engine
I didn't configured my database yet and realized that DatabaseError and
DatabaseModelError weren't been catched. Then, I realized that there is
already a lot of different errors to catch. To avoid an infinite list of
errors to think of, I just chose to catch at the highest level possible:
Exception.

It might probably look like a too generic catch, but we really don't
want to return PHP errors right in the front of the user. Also, the
error is still displayed with a proper http status code and in a HTML
document.

Later, when a log system will exist, I might add some warning/error logs
right here.

There's still some errors that I don't catch: if the error responses
fail, and if an action doesn't return a response. It will be done later.