tec: Allow empty Response view pointer
I was a bit annoyed by the fact of being forced to create a view file
for each controller action, even to return an empty answer (which is
perfectly fine). Also, I found weird to have different method signatures
for "ok" response and "error" ones. It is now the same for all: an empty
view pointer.