tec: Replace Response::fromCode by constructor
I realized that a developer might be tempted to return `new Response()`
from an action. Unfortunately, such a response is unusable and
parameters must be set. While improving the API of the constructor, I
realized that the `fromCode` method was actually a constructor.

Also, `fromCode` was mainly used internally, so there was really no
reasons to not transform it in a constructor.