tec: Add a simple template system to View
I wanted to reuse the template of the home page for my error pages, but
didn't want to duplicate the HTML boilerplate.

Adding a `template` method to set a template for the current view was
really easy and allow to use it from within an included template,
allowing quite powerful systems.