tec: Move autoload files in root directories
I plan to simplify the application architecture a more bit so it's
easier to identify purpose of files and directories. I also want to
avoid name conflict (for instance, at the moment we can't have a
`models` or a `views` controller because those directories are already
used for the application).

Following this logic, I want to avoid a name conflict with the
`autoload.php` file, this is why I moved the `src/` one in the root
directory. In order to be coherent, I did the same with the Minz one.