tec: Simplify application structure
It is some kind of revert of fa8a1de since controller view files are
moving back to src/views. The main difference is that controllers are
placed in src/name_of_controller.php files.

The previous structure wasn't as easy and intuitive as I expected. It
was quite difficult to distinguish controllers directories from other
"structural" directories such as `models` or `views`, and worse than
that: their names was conflicting.

Controllers are very central and are the entrypoints of the app, so I
think it's good to place them directly in src.

Also, the previous view pointer system didn't make sense anymore, so I
removed it while keeping the terminology. I have some ideas to make it
useful again!

This is a new try to make things better. It doesn't mean it's my final
decision ;)