Extract application initialization in a class
I'll need to reuse most of the application initialization from the CLI
file. Also it makes sense to initialize the router next to the engine,
in the src directory.

I'm not sure of what to do with the database initialization, but it's
easier for me in the Application for now.