Allow to build Subscription from values
We need to manipulate Subscription objects from the database. We can't
rely directly on the data since some logic needs to be handled by the
model. It makes sense to be able to create the model from the database
data, but it's also a lot of checks to be sure to initialize a correct
object.

Maybe a better approach would be to do no check at all. I may improve
all of that later.