Test controllers actions through Application
In previous tests, I had to initialize a Request with a method and a
path which weren't useful and might be wrong. Also, I didn't test that
the router was correctly configured, resulting in potential fix commits
(see 588ccc2).

I realized there was very little work to do to fix those two problems.
The Application class is doing exactly what I needed and is very easy to
initialize. Also, since it includes the controller file through the
Engine -> ActionController, I no longer need the `includeController`
helper method.

It allows me to test easily the full chain of the application, from
routing to the response.

I love this version of Minz <3