tec: Trim Migrator version number before setting
Some editors automatically add empty new line at the end of the files.
In the `data/migrations_version.txt`, this could lead to the situation
where the stored version is "<version_number>\n". We must trim this
string to avoid the system to fail. It could be done outside of the
Migrator class, but it would be too error-prone in my opinion.