fix: Fix property type validation
We use `filter_var` function to test the value of a property. This
function returns the casted value if the test is successful, false in
case of an error. Problem was that we didn't test correctly this value
and value "0" (zero) was considered as an error. Also, the "false" value
when checking a boolean must be tested differently since `filter_var`
can return it in case of success.