fix: Don't change secret and lease on renew
If an attacker tries to renew a subscription with a different secret
and/or lease, the intent would not be verified but the secret and lease
would change nevertheless.

Now, we store the values in new `pending_secret` and
`pending_lease_seconds` properties, waiting for the verification to set
real `secret` and `lease_seconds`.