tec: Always display errors from Engine
If not_found.phtml and/or internal_server_error.phtml view files were
not existing, the errors in Engine class were silently catched and it
was difficult for developers to understand what was going on.

It now relies on a new interface: Output. A View implements the Output,
such as the new Text class. It proposes to developers different
interfaces to display the content to the users.